+++ docs/release-notes.md
@@ -8,6 +8,8 @@ nav_order: 9
 
 ### Features
 
+- Add support for loading `inline` and `local` butane files, these are automatically translated to `ignition`.
+
 ### Bug fixes
 
 - Stop overriding default LUKS cipher algorithm in FIPS mode _(openshift 4.20.0-exp)_